requirements.txt
# PyTorch - Elige UNA de estas dos líneas según tu hardware
# torch==2.1.2 --index-url https://download.pytorch.org/whl/cpu  # Para CPU
torch==2.1.2 --index-url https://download.pytorch.org/whl/cu121  # Para GPU NVIDIA

# Transformers y Diffusers
transformers==4.35.2
diffusers==0.24.0

# Procesamiento de imágenes
Pillow==10.1.0
numpy==1.24.3

# Barra de progreso
tqdm==4.66.1

# Fine-tuning (LoRA)
peft==0.7.1

# Aceleración
accelerate==0.25.0

# Flask y web
flask==3.0.0
flask-restful==0.3.10
flask-cors==4.0.0
python-dotenv==1.0.0

# Utilidades
requests==2.31.0
safetensors==0.4.1